+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [Unreleased]
+### Updated
+- Now an empty object on creator details validator is allowed
+
 ## [1.7.0-rc.0] - 2019-05-21
 ### Added
 - Validators for specified types: